+++ apps/www/src/lib/mdx.ts
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import path from 'path'
 
-import { Option, Post } from '@/types'
+import type { Option, Post } from '@/types'
 
 const GENERATED_PATH = path.resolve(process.cwd(), 'generated')
 
@@ -51,5 +51,5 @@ export const getPosts = (): Post[] => {
 export const getPost = (slug: string): Option<Post> => {
   const allPosts = getPosts()
   const post = allPosts.find(post => post.slug === slug)
-  return post || null
+  return post ?? null
 }