+++ packages/usehooks-ts/src/useBoolean/useBoolean.test.ts
@@ -1,9 +1,9 @@
-import { act, renderHook } from '@testing-library/react-hooks/dom'
+import { act, renderHook } from '@testing-library/react'
 
 import { useBoolean } from './useBoolean'
 
 describe('useBoolean()', () => {
-  test('should use boolean', () => {
+  it('should use boolean', () => {
     const { result } = renderHook(() => useBoolean())
 
     expect(result.current.value).toBe(false)
@@ -13,40 +13,40 @@ describe('useBoolean()', () => {
     expect(typeof result.current.setValue).toBe('function')
   })
 
-  test('should default value works (1)', () => {
+  it('should default value works (1)', () => {
     const { result } = renderHook(() => useBoolean(true))
 
     expect(result.current.value).toBe(true)
   })
 
-  test('should default value works (2)', () => {
+  it('should default value works (2)', () => {
     const { result } = renderHook(() => useBoolean(false))
 
     expect(result.current.value).toBe(false)
   })
 
-  test('should set to true (1)', () => {
+  it('should set to true (1)', () => {
     const { result } = renderHook(() => useBoolean(false))
 
     act(() => {
-      result.current.setTrue()
-    })
-
-    expect(result.current.value).toBe(true)
-  })
-
-  test('should set to true (2)', () => {
-    const { result } = renderHook(() => useBoolean(false))
-
-    act(() => {
-      result.current.setTrue()
       result.current.setTrue()
     })
 
     expect(result.current.value).toBe(true)
   })
 
-  test('should set to false (1)', () => {
+  it('should set to true (2)', () => {
+    const { result } = renderHook(() => useBoolean(false))
+
+    act(() => {
+      result.current.setTrue()
+      result.current.setTrue()
+    })
+
+    expect(result.current.value).toBe(true)
+  })
+
+  it('should set to false (1)', () => {
     const { result } = renderHook(() => useBoolean(true))
 
     act(() => {
@@ -56,7 +56,7 @@ describe('useBoolean()', () => {
     expect(result.current.value).toBe(false)
   })
 
-  test('should set to false (2)', () => {
+  it('should set to false (2)', () => {
     const { result } = renderHook(() => useBoolean(true))
 
     act(() => {
@@ -67,7 +67,7 @@ describe('useBoolean()', () => {
     expect(result.current.value).toBe(false)
   })
 
-  test('should toggle value', () => {
+  it('should toggle value', () => {
     const { result } = renderHook(() => useBoolean(true))
 
     act(() => {
@@ -77,7 +77,7 @@ describe('useBoolean()', () => {
     expect(result.current.value).toBe(false)
   })
 
-  test('should toggle value from prev using setValue', () => {
+  it('should toggle value from prev using setValue', () => {
     const { result } = renderHook(() => useBoolean(true))
 
     act(() => {