+++ packages/usehooks-ts/src/useCopyToClipboard/useCopyToClipboard.test.ts
@@ -1,33 +1,33 @@
-import { act, renderHook } from '@testing-library/react-hooks/dom'
+import { act, renderHook } from '@testing-library/react'
 
 import { useCopyToClipboard } from './useCopyToClipboard'
 
-describe('useClipboard()', () => {
+describe('useCopyToClipboard()', () => {
   const originalClipboard = { ...global.navigator.clipboard }
   const mockData = 'Test value'
 
   beforeEach(() => {
     const mockClipboard = {
-      writeText: jest.fn(),
+      writeText: vitest.fn(),
     }
     // @ts-ignore mock clipboard
     global.navigator.clipboard = mockClipboard
   })
 
   afterEach(() => {
-    jest.resetAllMocks()
+    vitest.resetAllMocks()
     // @ts-ignore mock clipboard
     global.navigator.clipboard = originalClipboard
   })
 
-  test('should use clipboard', () => {
+  it('should use clipboard', () => {
     const { result } = renderHook(() => useCopyToClipboard())
 
     expect(result.current[0]).toBe(null)
     expect(typeof result.current[1]).toBe('function')
   })
 
-  test('should copy to the clipboard and the state', async () => {
+  it('should copy to the clipboard and the state', async () => {
     const { result } = renderHook(() => useCopyToClipboard())
 
     await act(async () => {