+++ packages/usehooks-ts/src/useElementSize/useElementSize.demo.tsx
@@ -1,12 +1,14 @@
 import { useState } from 'react'
 
-import { useElementSize } from '..'
+import { useElementSize } from './useElementSize'
 
 export default function Component() {
   const [isVisible, setVisible] = useState(true)
   const [squareRef, { width, height }] = useElementSize()
 
-  const toggleVisibility = () => setVisible(x => !x)
+  const toggleVisibility = () => {
+    setVisible(x => !x)
+  }
 
   return (
     <>