+++ packages/usehooks-ts/src/useTernaryDarkMode/useTernaryDarkMode.demo.tsx
@@ -1,4 +1,5 @@
-import { useTernaryDarkMode } from '..'
+import type { TernaryDarkMode } from './useTernaryDarkMode'
+import { useTernaryDarkMode } from './useTernaryDarkMode'
 
 export default function Component() {
   const {
@@ -7,7 +8,6 @@ export default function Component() {
     setTernaryDarkMode,
     toggleTernaryDarkMode,
   } = useTernaryDarkMode()
-  type TernaryDarkMode = typeof ternaryDarkMode
 
   return (
     <div>
@@ -24,9 +24,10 @@ export default function Component() {
         <br />
         <select
           name="select-ternaryDarkMode"
-          onChange={ev =>
+          onChange={ev => {
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
             setTernaryDarkMode(ev.target.value as TernaryDarkMode)
-          }
+          }}
           value={ternaryDarkMode}
         >
           <option value="light">light</option>