+++ packages/usehooks-ts/src/useWindowSize/useWindowSize.demo.tsx
@@ -1,12 +1,11 @@
-import { useWindowSize } from '..'
+import { useWindowSize } from './useWindowSize'
 
 export default function Component() {
-  const { width, height } = useWindowSize()
+  const size = useWindowSize()
 
   return (
     <div>
-      The current window dimensions are:{' '}
-      <code>{JSON.stringify({ width, height })}</code>
+      The current window dimensions are: <code>{JSON.stringify(size)}</code>
     </div>
   )
 }